fix(depgraph): force re-scan on hash drift to prevent stale-artifact hits (#449)
When the source file was journal-fresh but the recomputed artifact_key
differed from the entry's stored key, `check` / `check_diagnostic`
returned `Hit { artifact_key: K_predict }` derived from the STALE
`entry.resolved_includes` set captured by the previous `update()`.
The pipeline then looked K_predict up in the artifact store; a
coincidentally matching older artifact (from an intermediate state
whose byte-stream set happened to hash to the same key) could be
served as a stale `.obj` — manifesting as undefined-symbol link
errors after a `.cpp.hpp` unity-build header was modified.

Use `last_file_hashes` as a content-drift signal independent of the
journal: any tracked path whose current hash differs from the stored
one — or any path that wasn't in `last_file_hashes` to begin with —
forces a `HeadersChanged` verdict so the pipeline recompiles,
re-scans the dependency set, and re-stores under the write-side key
derived from the post-compile inputs.

Adds `drifted_paths` + `format_drift_for_log` helpers shared by the
fast (`check`) and diagnostic (`check_diagnostic`) verdict paths.
Splits `depgraph/graph.rs` into `graph/mod.rs` + `graph/tests.rs` so
the file no longer hard-blocks edits at the LOC guard. Adds
`tests/depgraph_drift_detection_test.rs` with three regression
guards: drifted-header, drifted-source, and unchanged-warm.
